Prevent deletion of Service Type if assigned to any vehicles.
[Finishes #55764018]

+++ app/services/service_type_service.rb
@@ -17,9 +17,9 @@ class ServiceTypeService
   end
 
   def delete_service_type(service_type)
-    #if service_type.services.any?
-    #  return :service_exists
-    #end
+    if service_type.services.any?
+      return :service_exists
+    end
 
     service_type.destroy
   end
